+++ packages/plugin-video/web/src/toolbar/videoSelectionInputModal.jsx
@@ -12,11 +12,14 @@ export default class VideoSelectionInputModal extends Component {
     const { componentData } = this.props;
     this.state = {
       url: (!componentData.isCustomVideo && componentData.src) || '',
-      errorMsg: '',
     };
     this.id = `VideoUploadModal_FileInput_${Math.floor(Math.random() * 9999)}`;
-    const { onConfirm, onReplace } = this.props;
-    this.onConfirm = onConfirm || onReplace;
+    const onConfirm = props.onConfirm || props.onReplace;
+
+    this.onConfirm = args => {
+      this.setError(false);
+      return onConfirm(args);
+    };
   }
 
   onUrlChange = e => {
@@ -76,6 +79,10 @@ export default class VideoSelectionInputModal extends Component {
     this.onConfirm({ ...componentData, src, isCustomVideo });
   };
 
+  setError(error) {
+    this.props.pubsub.update('componentState', { error });
+  }
+
   setComponentData = data => {
     this.props.pubsub.set('componentData', data);
   };
@@ -88,14 +95,18 @@ export default class VideoSelectionInputModal extends Component {
     const { componentData, handleFileUpload: consumerHandleFileUpload } = this.props;
     const file = this.inputFile.files[0];
     this.loadLocalVideo(file);
-    consumerHandleFileUpload(file, ({ data, error }) =>
-      this.updateVideoComponent({ data, error }, componentData, true)
-    );
+    consumerHandleFileUpload(file, ({ data, error }) => {
+      if (error) {
+        this.setError(error);
+      } else {
+        this.updateVideoComponent({ data }, componentData, true);
+      }
+    });
     this.closeModal();
   };
 
   render() {
-    const { url, showError, errorMsg } = this.state;
+    const { url, showError } = this.state;
     const {
       t,
       handleFileSelection,
@@ -142,7 +153,6 @@ export default class VideoSelectionInputModal extends Component {
           >
             + {t('VideoUploadModal_CustomVideoClickText')}
           </label>
-          {errorMsg.length > 0 && <div className={styles.video_modal_error_msg}>{errorMsg}</div>}
         </div>
       </div>
     );