+++ packages/plugin-video/web/src/video-component.jsx
@@ -40,6 +40,18 @@ class VideoComponent extends React.Component {
     this.handlePlayerFocus();
   }
 
+  static getDerivedStateFromProps(props, state) {
+    const {
+      componentState: { error },
+    } = props;
+    if (error) {
+      return { errorMsg: error.msg };
+    } else if (error === false) {
+      return { errorMsg: undefined };
+    }
+    return {};
+  }
+
   handlePlayerFocus() {
     // eslint-disable-next-line react/no-find-dom-node
     const element = findDOMNode(this).querySelector('iframe, video');
@@ -111,7 +123,7 @@ class VideoComponent extends React.Component {
   render() {
     this.styles = this.styles || mergeStyles({ styles, theme: this.props.theme });
     const { className, onClick } = this.props;
-    const { isPlayable } = this.state;
+    const { isPlayable, isLoaded, errorMsg } = this.state;
     const containerClassNames = classNames(this.styles.video_container, className || '');
     /* eslint-disable jsx-a11y/no-static-element-interactions */
     return (
@@ -124,7 +136,8 @@ class VideoComponent extends React.Component {
       >
         {!isPlayable && this.renderOverlay(this.styles, this.props.t)}
         {this.renderPlayer()}
-        {!this.state.isLoaded && this.renderLoader()}
+        {!isLoaded && !errorMsg && this.renderLoader()}
+        {errorMsg && <div className={styles.error}>{errorMsg}</div>}
       </div>
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */